+++ packages/utils/src/internals/iterables.ts
@@ -1,3 +1,49 @@
+/**
+ * Type guard that checks if a value is iterable (has Symbol.iterator).
+ * @internal
+ *
+ * **Example usage:**
+ * ```ts
+ * if (isIterable(someValue)) {
+ *   for (const item of someValue) {
+ *     console.log(item);
+ *   }
+ * }
+ * ```
+ */
+export function isIterable<T>(value: unknown): value is Iterable<T> {
+    if (value == null || typeof value === 'string' || ArrayBuffer.isView(value)) {
+        return false;
+    }
+
+    if (Array.isArray(value)) {
+        return true;
+    }
+
+    return typeof Object(value)[Symbol.iterator] === 'function';
+}
+
+/**
+ * Type guard that checks if a value is async iterable (has Symbol.asyncIterator).
+ * @internal
+ *
+ * **Example usage:**
+ * ```ts
+ * if (isAsyncIterable(someValue)) {
+ *   for await (const item of someValue) {
+ *     console.log(item);
+ *   }
+ * }
+ * ```
+ */
+export function isAsyncIterable<T>(value: unknown): value is AsyncIterable<T> {
+    if (value == null || typeof value === 'string' || ArrayBuffer.isView(value)) {
+        return false;
+    }
+
+    return typeof Object(value)[Symbol.asyncIterator] === 'function';
+}
+
 // Source - https://stackoverflow.com/a/71288323
 /**
  * Merges multiple async iterables into a single async iterable, yielding values concurrently.
